requirements.txt
playwright==1.37.0
yt-dlp
aiogram
dotenv-python